fixes for srcdir != builddir
Note: kview looks as it could use USE_AUTOMOC, but then kmicoapp.h
is passed through moc and this bombs out! I think (I often think ;),
that this file should be removed out of the kview dir when it isn't
used there. Create for example kview/experiments or something to place
it there

svn path=/trunk/kdegraphics/kdvi/; revision=15702