Port of revision 470668
This simplifies the recently introduced toolsystem, and allows to easily add new tools.
The active tool is now also stored between sessions.

This also adds a measuring tool, which allows to measure position, distance and angle
in screen coordinates, as well as a variety of physical units like cm, inch, TeXPoints, ...

svn path=/trunk/KDE/kdegraphics/kviewshell/plugins/dvi/; revision=470681