Add support for multithreading to achieve a nonblocking GUI.
At the moment only the fax and djvu plugins use this feature.
For the pdf plugin a small change to libpoppler is needed first.
And since the dvi plugin uses QPainter to draw the pages using
a thread for page rendering is not possible without major changes.
Qt4 should probably help.

There are still a few little issues, but I commit this now since the patch
is already quite large.

svn path=/branches/work/kviewshell-0.7/kviewshell/plugins/dvi/; revision=493048