Add multithreading to the DVI plugin. This one is not as smooth as the rest of the plugins, because it uses QPainter for the drawing. And since X is not multithreaded we therefore need to lock the whole event loop before each drawing operation.
In Qt4 this can be avoided by simple drawing onto a QImage.

There are a few other problems: For now only DVI files that contain a paper size special
work correctly, and all warning messages from the renderer have been disabled, since
we cannot just open a KMessageBox in a thread.

Still, at last we can now remove the single threaded API, which should allow for quite a
few simplifications in the code.

svn path=/branches/work/kviewshell-0.7/kviewshell/plugins/dvi/; revision=514021