Backport from SVN commit 619673 by pino in okular:
Fix a nasty bug that could happened in presentation mode:
when requestng -synchronously- a pixmap for the current page, it could be scheduled in the queue. Then, requesting the previous and next page in presentation mode -in a second request- caused the deletion of the original page request.
The fix consists in requesting all the requests -at once-, so no request is lost.

svn path=/branches/KDE/3.5/kdegraphics/kpdf/; revision=619946