Backport from SVN commit 619810 by pino in okular:
One single char remove to create correctly the queue of pixmap requests: new asynchronous requests goes as last one in the group of other with the same priority.
Thus, if an observer requests eg pixmap for pages 2 and 3, now they are generated in the correct order the observer specified, and not in the reversed one.

svn path=/branches/KDE/3.5/kdegraphics/kpdf/; revision=619947