Add support for encrypted ODF documents.
This is a bit experimental, but it does (mostly :-)) work for me.

Note that the KWallet integration should be interoperable with
KWord, so if you enter the password in one application, you
shouldn't be prompted for the password in the other (as long as
the document path is the same - it won't work if you move the 
document).

svn path=/trunk/KDE/kdegraphics/okular/; revision=774836