Backport:
Make the presentation mode a normal widget (instead of a "fake" dialog), so it is not treated like a real dialog when compositing is active.
CCBUG: 163013

While I am on it, disable the active modality of the presentation mode, so the main window can be actively used once a presentation is active.
CCBUG: 152585

svn path=/branches/KDE/4.0/kdegraphics/okular/; revision=817704