Fix Print to File output by using Ghostscript ps2pdf to convert ps print file into pdf format and save that.  Only enabled in PDF, DVI, and DJVU generators if ps2pdf is found in the path.  Ghostscript is almost always guaranteed to be installed given many parts of the printing stack depends on it.
Note that resulting PDF or PS file will have the page size as set in the print
dialog and not the page size of the original document, i.e. if an A5 booklet
is printed but the dialog page size is set to A4 then the resulting file will
be A4 and not A5.

svn path=/trunk/KDE/kdegraphics/okular/; revision=897000