When loading a comicbook, try to get just the size of the various images without loading them if possible.
Instead of just fully loading the images, make use of QImageReader to query for the size if available, and in case just do the old way (full image loading).
This drastically kills the loading time of comicbook documents of at least 50%, even by 75% (and possibly something more) on the average case.

svn path=/trunk/KDE/kdegraphics/okular/; revision=917041