Apply a way to resolve the real location of resources to load.
When iterating through the document HTML files, set in the epub document the URL of the current file being loaded,
so it is possible to resolve the locations of resources based on the base URL of the current file.
This makes it possible to load resources like "foo.css", "somedir/foo.css", "../somedir/foo.css" and so on.

BUG: 198764
FIXED-IN: 4.6

svn path=/trunk/KDE/kdegraphics/okular/; revision=1182629