Allow to specify which config file the part uses.
This is done through a string "ConfigFileName=<file name>" which can be given to
the constructor in its argument list.

Note that the file ui/fileinterpreterpreview.cpp currently still contains a
reference to "okularrc", which we leave for now (as it just stores the size
of a dialog in that config file).