Fixes in presentation page switch logic
Do not assume that oldPage == m_document->viewport().pageNumber, because
it's false if we are processing a viewport change notification.
It caused presentation drawings to turn into persistent annotations,
when switching page from the main okular window.

In general, this patch also tidies up PresentationWidget::changePage a bit:
 - Removed redundant checks: they are always true
 - Removed duplicated code that performed the page opening actions:
     Previously the first slide and the other ones followed different code
     paths, both doing the same things. I unified them.

I also changed slotNextPage so that it never sets m_frameIndex = -1, because
it confuses changePage and makes it skip page closing actions.