Removed calls to query size of pixmap size in PickPointEngine::event
Since the size of pixmap is determined by the value of size variable,
we can directly read that variable instead of querying the pixmap.
Also note that "pixmap" comes from "hoverIcon", not from "icon", and
that piece of code wants to know the size of the latter. In case of
stamp annotations, however, they happen to be the same.