Do not delete the Part on the Shell destructor
The machinery in KParts/QObject is already doing it and this way we don't get the
KXMLGUIClient::~KXMLGUIClient: 0x15637528 deleted without having been removed from the factory first. This will leak standalone popupmenus and could lead to crashes.
warning

I'm not sure if calling this a kdelibs bug yet or not though :D

BUGS: 261538
FIXED-IN: 4.9.1