Fix annotation positioning when using tiles
Most of the positioning calculations were taking into account the 'crop'
argument (which is a NormalizedRect) to execute some translations.

If not using tiles the value of crop were always the normalized bounding
box of the page, relative to the page itself. So translating by it is
essentially a noop.

When using tiles this argument represents the normalized viewport,
relative to the page. Although useful for the tiles manager, translating
by 'crop' is not necessary for any of those operations.