Fix word detection for DVI documents
This patch attempts to restore the functionalities broken by some changes
(maybe the text column selection code, it was broken also in 4.7).

Text search and text selection work (almost) properly again.

It uses a bit of heuristics to identify the end of a word and merge the
boxes which enclose each character of a word (so that
char_x.right==char_{x+1}.left).
It also tries to recognize if there is a newline ("after_space") after
that a space is found.

BUG: 302004
REVIEW: 107429
FIXED-IN: 4.9.4