Get rid of window background-colored bars between thumbnail containers.
This seems to be what the original code wanted to achieve, but failed
due to a convoluted approach of setting the color of the current back-
ground role to a default palette's Base color ... instead we just leave
the palette untouched and set the correct background role now. Visual
noise be gone, yay.

CCMAIL:wstephenson@kde.org