Fix the issue exposed by the previous patch
Cherry-picked from a556126816875b7e2a4da3924d400e1c8bce6530

The patch de-singletons the PageController class.

The PageController is now per-document and it gets deleted when the
document is closed.

As consequence of this, the RotationJob's done signal will not be
delivered if the document has been closed, and thus this fixes the
crash.