Improve XY Cut layout recognition code
It was a simple bug in the XY Cut layout recognition code that made it too eager to see columns everywhere.
Also removed the dependence of the layout analysis algorithms on the display DPI (introduced by the recently added feature of using KScreen) to make their behavior more predictable and reproducible.

BUGS: 326207
BUGS: 331090
FIXED-IN: 4.13.0
REVIEW: 115759