Modify how page size label is shown on the bottom bar
Opened a file with all pages with same size and behaves exactly like Okular does
Opened a file with pages with different sizes and behaves like Okular does until it stops fiting. In that moment the page bar is not centered anymore to let the text still fit and once the text doesn't fit anymore it starts getting ellided.

Reviewed by Luigi

BUGS: 333349
FIXED-IN: 4.13.1