Make KActivities a required dependency
Our goal is better integration with Plasma and other KDE components.

We have made a conscious decision to reduce our code a bit by removing
the legacy #ifdefs. That way we do not have to support a build
configuration which is going to be used by just a tiny minority of our
users, and one that we are not terribly interested in. Without
KActivities, there is for example no support for recording and listing
of recent documents, PDF sharing etc etc, and we honestly believe that
the majority of our audience wants these features.

REVIEW: 123249