Fix play back of video with rendition link
Process the operations of the rendition link independently of whether
there is a JS script or not. Most of the time the JS script won't be
executed correctly anyway, since Okular is missing most of the JS API,
so we should fall back on the given operations.

BUGS: 324167
FIXED-IN: KDE Applications 15.08.0