Fix picking of different view modes
This was broken since commit 077689c4 when this got ported away from
KAction. The code attempted to add a parent QMenu to the QAction
instances nested below that parent. I'm quite surprised that it didn't
crash, but checkable actions do not go well with actions that have
nested menus, apparently.

I haven't checked the rest of the code for similar mistakes, but I don't
see any other case of this bug in Okular's menus when reading PDFs.

REVIEW: 124842