Set objectName for 'm_name'
The unit test 'editdrawingtooldialogtest' on the frameworks branch aborts at run-time.  Reason: it expects the dialog to have a KLineEdit member with object name 'name'.  That KLineEdit does exist in the dialog, but no object name is set for it at all.  This patch explicitly sets the object name to 'name', which makes the editdrawingtooldialogtest pass again.

REVIEW: 126158