Bug 236688 - Sort pages according to TOC in CHM Files
All pages in a CHM File are now ordered accordingly to TOC Data. If a page is not in TOC, it will be added at the end.

Testing Done:
I created a small example chm file to test this. I can provide the file, if someone would like to have it.

Bugs closed: 236688

Reviewed at https://git.reviewboard.kde.org/r/130078/