Implement rasterized printing with QPrinter with hidden annotations
Previously, rasterized printing would use the QPrinter backend whenever
annotations were to be printed, and fall back to the convert-to-ps
toolchain otherwise.  Effectively, this meant that printing on
windows without showing the annotations was not possible (because
on windows only the QPrinter can be used for printing).

Starting with poppler 0.60, there is a way to disable annotation
rendering for the renderToImage and renderToPainter methods that
are used by QPrinter.  This patch makes use of this new option,
making printing on windows without annotations possible.

Differential Revision: https://phabricator.kde.org/D7688