Add signature panel
Summary:
Signature panel is a sidebar widget that lists all the signatures present in the
document. The information will be available in a tree like structure. Also clicking
on certain rows will change the document viewport to where the signature is located
and breifly change the color of the signature form widget.

ss:
{F6191765}

Depends on D13366

Reviewers: aacid

Subscribers: okular-devel

Tags: #okular

Differential Revision: https://phabricator.kde.org/D14816