Fix page boundary gradients on scaled displays
The code that draws a thin gradient at the right and bottom
boundaries of a page implicitly assumed that the display was
not scaled, i.e., that logical pixel where identical to
physical pixels.  Nobody really noticed because that gradient
was so thin.  Anyway, this patch fixes the gradient for
scaled displays.

I considered using a QLinearGradient, but decided against it:
The code would have been longer, and the manual code in the
patch can exploit the fact that all lines are horizontal
or vertical.  Therefore, expensive scan-conversion can be
avoided.