Add kinetic scrolling & smooth navigation with cursor keys
This patch adds kinetic / inertial scrolling to ease the use of Okular on touchscreen devices, using Qt's QScroller. I think it's much nicer even for mouse usage.

I left the "mouse cursor wraps when going at the edge of the screen" mechanism intact.

Also, I made it so that there is a nice animation when using the up / down keys to navigate the document, instead of having discrete steps.

FEATURE: 413989

FIXED-IN: 1.10.0