Improve the sidebar's navigation and UX
Okular's sidebar vertical view chooser toolbar suffers from a few issues:

* It's a nonstandard UI not used for category choosers in other pieces of KDE software,
  and not used in other FOSS document readers
* What is shown and what is hidden is simultaneously too configurable while still not
  offering the desirable UI common to other programs (i.e. no visible category chooser,
  but a sidebar capable of displaying thumbnails, table of contents, search results, etc.)
* With labels on it takes up quite a bit of horizontal space, while with labels off, the
  categories are less than clear
* UX is kind of clunky with nonstandard behaviors (e.g. clicking on the current category
  to hide that category's view while keeping the view chooser visible, showing mostly
  disabled items)
* It's made with custom painting code, which reduces maintainability and introduces bugs
  (e.g. https://bugs.kde.org/show_bug.cgi?id=408190)

This patch removes the vertical category chooser entirely and replaces it with a tabbed
view on the top of the sidebar itself. The tabs are icons-only and have large icons. A
button is added on the left side of the default toolbar to quickly hide or show the
sidebar. In order to make room for the new button, the Previous and Next buttons on the
toolbar are removed, as previous/next buttons are already present on the Page Bar on the
bottom of the window so there's no need to duplicate this functionality. This improves
the UX, fixes a variety of bugs, and deletes a lot of custom code of dubious long-term
maintainability.

![vokoscreenNG-2020-04-16_13-29-24](https://invent.kde.org/graphics/okular/uploads/a1f96a315b69282df51de9993b1befaf/vokoscreenNG-2020-04-16_13-29-24.webm)

BUG: 213508

BUG: 334441

BUG: 344599

BUG: 408190

CCBUG: 335189

FIXED-IN: 1.11.0

CHANGELOG: The sidebar can now be easily shown or hidden with a toolbar button, and the category chooser no longer takes up so much space